chore: Move CSS patch for ipywidget in VSCode in its module
